+++ hpvsim/hiv.py
@@ -16,25 +16,65 @@ class HIVsim(hpb.ParsObj):
         A class based around performing operations on a self.pars dict.
         '''
 
-    def __init__(self, sim, art_datafile, hiv_datafile, hiv_pars=None):
+    def __init__(self, sim, art_datafile, hiv_datafile, hiv_pars):
         pars = self.load_data(hiv_datafile=hiv_datafile, art_datafile=art_datafile)
         self.people = sim.people
         # Define default parameters, can be overwritten by hiv_pars
         pars['hiv_pars'] = {
+<<<<<<< HEAD
             'rel_sus': 2.2,  # Increased risk of acquiring HPV
             'rel_hiv_sev': {'cd4_200': 1.2, 'cd4_200_500': 1.1},  # Relative severity for HIV
             'reactivation_prob': 3, # Unused for now, TODO: add in rel_reactivation to make functional
+=======
+            'rel_sus': { # Increased risk of acquiring HPV
+                'cat1': {
+                    'cd4_lower': 0,
+                    'cd4_upper': 200,
+                    'value': 2.2
+                },
+                'cat2': {
+                    'cd4_lower': 200,
+                    'cd4_upper': 500,
+                    'value': 2.2
+                }
+            },
+            'rel_sev_infl': { # Speed up growth of disease severity
+                'cat1':{
+                    'cd4_lower': 0,
+                    'cd4_upper': 200,
+                    'value': 0.25
+                },
+                'cat2': {
+                    'cd4_lower': 200,
+                    'cd4_upper': 500,
+                    'value': 0.5
+                }
+            },
+            'rel_imm': { # Reduction in neutralizing/t-cell immunity acquired after infection/vaccination
+                'cat1': {
+                    'cd4_lower': 0,
+                    'cd4_upper': 200,
+                    'value': 0.36
+                },
+                'cat2': {
+                    'cd4_lower': 200,
+                    'cd4_upper': 500,
+                    'value': 0.76
+                }
+            },
+            'rel_reactivation_prob': 3, # Unused for now, TODO: add in rel_reactivation to make functional
+>>>>>>> hiv-updates
             'time_to_hiv_death_shape': 2, # shape parameter for weibull distribution, based on https://royalsocietypublishing.org/action/downloadSupplement?doi=10.1098%2Frsif.2013.0613&file=rsif20130613supp1.pdf
             'time_to_hiv_death_scale': lambda a: 21.182 - 0.2717*a, # scale parameter for weibull distribution, based on https://royalsocietypublishing.org/action/downloadSupplement?doi=10.1098%2Frsif.2013.0613&file=rsif20130613supp1.pdf
             'cd4_start': dict(dist='normal', par1=594, par2=20),
             'cd4_trajectory': lambda f: (24.363 - 16.672*f)**2, # based on https://docs.idmod.org/projects/emod-hiv/en/latest/hiv-model-healthcare-systems.html?highlight=art#art-s-impact-on-cd4-count
-            'cd4_reconstitution': lambda m: 15.584*m - 0.2113*m**2 # growth in CD4 count following ART initiation
+            'cd4_reconstitution': lambda m: 15.584*m - 0.2113*m**2, # growth in CD4 count following ART initiation
+            'art_failure_prob': 0.1, # Percentage of people on ART who will not suppress virus successfully
+            'dt_art': 5.0 # Timestep at which people originally not on ART can initiate care
         }
 
         self.init_states()
-        if hiv_pars is not None:
-            pars['hiv_pars'] = sc.mergedicts(pars['hiv_pars'], hiv_pars)
-        self.update_pars(pars, create=True)
+        self.update_pars(old_pars=pars, new_pars=hiv_pars, create=True)
         self.init_results(sim)
 
         y = np.linspace(0,1,101)
@@ -43,11 +83,29 @@ class HIVsim(hpb.ParsObj):
 
         return
 
+    def update_pars(self, old_pars=None, new_pars=None, create=True):
+        if len(new_pars):
+            for parkey, parval in new_pars.items():
+                if isinstance(parval, dict):
+                    for parvalkey, parvalval in parval.items():
+                        if isinstance(parvalval, dict):
+                            for parvalkeyval, parvalvalval in parvalval.items():
+                                old_pars['hiv_pars'][parkey][parvalkey][parvalkeyval] = parvalvalval
+                        else:
+                            old_pars['hiv_pars'][parkey][parvalkey] = parvalval
+                else:
+                    old_pars['hiv_pars'][parkey] = parval
+
+        # Call update_pars() for ParsObj
+        super().update_pars(pars=old_pars, create=create)
+
 
     def init_states(self):
         hiv_states = [
             hpd.State('cd4', hpd.default_float, np.nan),
+            hpd.State('hiv', bool, False),
             hpd.State('art', bool, False),
+            hpd.State('date_hiv', hpd.default_float, np.nan),
             hpd.State('date_art', hpd.default_float, np.nan),
             hpd.State('date_dead_hiv', hpd.default_float, np.nan),
             hpd.State('dead_hiv', bool, False),
@@ -69,24 +127,45 @@ class HIVsim(hpb.ParsObj):
 
         na = len(sim['age_bins']) - 1  # Number of age bins
 
+        stock_colors = [i for i in set(hpd.stock_colors) if i is not None]
+
         results['hiv_infections'] = init_res('Number HIV infections')
-        results['hiv_infections_by_age'] = init_res('Number HIV infections by age', n_rows=na)
-        results['n_hiv'] = init_res('Number living with HIV')
-        results['n_hiv_by_age'] = init_res('Number living with HIV by age', n_rows=na)
-        results['hiv_prevalence'] = init_res('HIV prevalence')
-        results['hiv_prevalence_by_age'] = init_res('HIV prevalence by age', n_rows=na)
-        results['hiv_incidence'] = init_res('HIV incidence')
-        results['hiv_incidence_by_age'] = init_res('HIV incidence by age', n_rows=na)
+        results['hiv_infections_by_age'] = init_res('Number HIV infections by age', n_rows=na, color=stock_colors[0])
+        results['n_hiv'] = init_res('Number living with HIV', color=stock_colors[0])
+        results['n_hiv_by_age'] = init_res('Number living with HIV by age', n_rows=na, color=stock_colors[0])
+        results['hiv_prevalence'] = init_res('HIV prevalence', color=stock_colors[0])
+        results['hiv_prevalence_by_age'] = init_res('HIV prevalence by age', n_rows=na, color=stock_colors[0])
+        results['hiv_incidence'] = init_res('HIV incidence', color=stock_colors[0])
+        results['hiv_incidence_by_age'] = init_res('HIV incidence by age', n_rows=na, color=stock_colors[0])
+        results['n_hpv_by_age_with_hiv'] = init_res('Number HPV infections by age among HIV+', n_rows=na, color=stock_colors[0])
+        results['n_hpv_by_age_no_hiv'] = init_res('Number HPV infections by age among HIV-', n_rows=na, color=stock_colors[0])
+        results['hpv_prevalence_by_age_with_hiv'] = init_res('HPV prevalence by age among HIV+', n_rows=na, color=stock_colors[0])
+        results['hpv_prevalence_by_age_no_hiv'] = init_res('HPV prevalence by age among HIV-', n_rows=na, color=stock_colors[1])
+        results['cancers_by_age_with_hiv'] = init_res('Cancers by age among HIV+', n_rows=na, color=stock_colors[0])
+        results['cancers_by_age_no_hiv'] = init_res('Cancers by age among HIV-', n_rows=na, color=stock_colors[1])
+        results['cancers_with_hiv'] = init_res('Cancers among HIV+', color=stock_colors[0])
+        results['cancers_no_hiv'] = init_res('Cancers among HIV-', color=stock_colors[1])
+        results['cancer_incidence_with_hiv'] = init_res('Cancer incidence among HIV+', color=stock_colors[0])
+        results['cancer_incidence_no_hiv'] = init_res('Cancer incidence among HIV-', color=stock_colors[1])
+        results['n_females_with_hiv_alive_by_age'] = init_res('Number females with HIV alive by age', n_rows=na)
+        results['n_females_no_hiv_alive_by_age'] = init_res('Number females without HIV alive by age', n_rows=na)
+        results['n_females_with_hiv_alive'] = init_res('Number females with HIV alive')
+        results['n_females_no_hiv_alive'] = init_res('Number females without HIV alive')
+        results['n_art'] = init_res('Number on ART')
+        results['art_coverage'] = init_res('ART coverage')
+
+
         self.results = results
         return
 
 
     # %% HIV methods
 
-    def set_hiv_prognoses(self, inds, year=None):
-        ''' Set HIV outcomes (for now only ART) '''
+    def set_hiv_prognoses(self, inds, year=None, incident=True):
+        ''' Set HIV outcomes '''
 
         art_cov = self['art_adherence']  # Shorten
+        shape = self['hiv_pars']['time_to_hiv_death_shape']
 
         # Extract index of current year
         all_years = np.array(list(art_cov.keys()))
@@ -99,20 +178,40 @@ class HIVsim(hpb.ParsObj):
         art_probs = np.zeros(len(self.people), dtype=hpd.default_float)
         art_probs[inds] = art_covs
 
+<<<<<<< HEAD
         # Get indices of people on ART
         art_inds = hpu.true(hpu.binomial_arr(art_probs))
 
+=======
+        # Get indices of people who are on ART
+        art_bools = hpu.binomial_arr(art_probs)
+        art_inds = hpu.true(art_bools)
+>>>>>>> hiv-updates
         self.people.art[art_inds] = True
         self.people.date_art[art_inds] = self.people.t
-        self.people.cd4[inds] = hpu.sample(**self['hiv_pars']['cd4_start'], size=len(inds))
+        self.people.date_dead_hiv[art_inds] = np.nan
+        self.people.dur_hiv[art_inds] = np.nan
 
-        # Draw time to HIV mortality
-        shape = self['hiv_pars']['time_to_hiv_death_shape']
-        scale = self['hiv_pars']['time_to_hiv_death_scale'](self.people.age[inds])
+        if incident:
+            # Filter those who are not on ART and assign time to HIV death
+            no_art_inds = np.setdiff1d(inds, art_inds)
+            scale = self['hiv_pars']['time_to_hiv_death_scale'](self.people.age[no_art_inds])
+            scale = np.maximum(scale, 0)
+            time_to_hiv_death = weibull_min.rvs(c=shape, scale=scale, size=len(no_art_inds))
+            self.people.dur_hiv[no_art_inds] = time_to_hiv_death
+            self.people.date_dead_hiv[no_art_inds] = self.people.t + sc.randround(time_to_hiv_death / self.people.dt)
+
+        # Find those on ART who will not be virologically suppressed and assign time to HIV death
+        art_failure_prob = self['hiv_pars']['art_failure_prob']
+        art_failure_probs = np.full(len(art_inds), fill_value=art_failure_prob, dtype=hpd.default_float)
+        art_failure_bools = hpu.binomial_arr(art_failure_probs)
+        art_failure_inds = art_inds[art_failure_bools]
+
+        scale = self['hiv_pars']['time_to_hiv_death_scale'](self.people.age[art_failure_inds])
         scale = np.maximum(scale, 0)
-        time_to_hiv_death = weibull_min.rvs(c=shape, scale=scale, size=len(inds))
-        self.people.dur_hiv[inds] = time_to_hiv_death
-        self.people.date_dead_hiv[inds] = self.people.t + sc.randround(time_to_hiv_death / self.people.dt)
+        time_to_hiv_death = weibull_min.rvs(c=shape, scale=scale, size=len(art_failure_inds))
+        self.people.dur_hiv[art_failure_inds] = time_to_hiv_death
+        self.people.date_dead_hiv[art_failure_inds] = self.people.t + sc.randround(time_to_hiv_death / self.people.dt)
 
         return
 
@@ -124,6 +223,7 @@ class HIVsim(hpb.ParsObj):
         filter_inds = self.people.true('hiv')
         inds = self.people.check_inds(self.people.dead_hiv, self.people.date_dead_hiv, filter_inds=filter_inds)
         self.people.remove_people(inds, cause='hiv')
+        self.people['hiv'][inds] = False
         return
 
 
@@ -137,15 +237,18 @@ class HIVsim(hpb.ParsObj):
             not_art_inds = filter_inds[hpu.false(self.people.art[filter_inds])]
 
             # First take care of people not on ART
-            frac_prognosis = 100*(self.people.t - self.people.date_hiv[not_art_inds])* self.people.dt/self.people.dur_hiv[not_art_inds]
+            frac_prognosis = 100 * (self.people.t - self.people.date_hiv[not_art_inds]) * self.people.dt / \
+                             self.people.dur_hiv[not_art_inds]
             cd4_change = self.cd4_decline_diff[frac_prognosis.astype(np.int64)]
             self.people.cd4[not_art_inds] += cd4_change
 
             # Now take care of people on ART
-            months_on_ART = (self.people.t - self.people.date_art[art_inds])*12
+            mpy = 12
+            months_on_ART = (self.people.t - self.people.date_art[art_inds]) * mpy
             cd4_change = self['hiv_pars']['cd4_reconstitution'](months_on_ART)
             self.people.cd4[art_inds] += cd4_change
 
+<<<<<<< HEAD
             cd4_200_inds = sc.findinds(self.people.cd4 < 200)
             cd4_200_500_inds = sc.findinds((self.people.cd4 > 200) & (self.people.cd4 < 500))
 
@@ -156,6 +259,19 @@ class HIVsim(hpb.ParsObj):
             if len(cd4_200_500_inds):
                 self.people.rel_sev[cd4_200_500_inds] = self['hiv_pars']['rel_hiv_sev']['cd4_200_500']
                 self.people.rel_sus[cd4_200_500_inds] = self['hiv_pars']['rel_sus']
+=======
+            inds_to_update = sc.autolist()
+            for ir, rel_par in enumerate(['rel_sus', 'rel_sev_infl', 'rel_imm']):
+                for cat, catvals in self['hiv_pars'][rel_par].items():
+                    inds = sc.findinds((self.people.cd4 >= catvals['cd4_lower']) & (self.people.cd4 < catvals['cd4_upper']))
+                    if len(inds):
+                        inds_to_update += list(inds)
+                        self.people[rel_par][inds] = catvals['value']
+
+            if len(inds_to_update):
+                inds_to_update = np.array(list(set(inds_to_update)))
+                self.update_hpv_progs(inds_to_update)
+>>>>>>> hiv-updates
 
         return
 
@@ -164,13 +280,23 @@ class HIVsim(hpb.ParsObj):
         '''
         Wrapper method that checks for new HIV infections, updates prognoses, etc.
         '''
+        # Pull out anyone with prevalent infection who is not on ART, check if they get on today
+        t = self.people.t
+        dt = self.people.dt
+
+        update_freq = max(1, int(self['hiv_pars']['dt_art'] / dt)) # Ensure it's an integer not smaller than 1
+        if t % update_freq == 0:
+            no_art_inds = hpu.true(self.people.hiv * ~self.people.art)
+            if len(no_art_inds):
+                self.set_hiv_prognoses(no_art_inds, year=year, incident=False)
 
         new_infection_inds = self.new_hiv_infections(year) # Newly acquired HIV infections
         if len(new_infection_inds):
             self.set_hiv_prognoses(new_infection_inds, year=year)  # Set ART adherence for those with HIV
-            self.update_hpv_progs(new_infection_inds) # Update any HPV prognoses
+
         self.check_hiv_mortality()
         self.check_cd4()
+        self.update_hiv_results(new_infection_inds)
         new_infections = self.people.scale_flows(new_infection_inds) # Return scaled number of infections
         return new_infections
 
@@ -198,17 +324,8 @@ class HIVsim(hpb.ParsObj):
         # Get indices of people who acquire HIV
         hiv_inds = hpu.true(hpu.binomial_arr(hiv_probs))
         self.people.hiv[hiv_inds] = True
-        self.people.hiv[hiv_inds] = self['hiv_pars']['cd4_start']
+        self.people.cd4[hiv_inds] = hpu.sample(**self['hiv_pars']['cd4_start'], size=len(hiv_inds))
         self.people.date_hiv[hiv_inds] = self.people.t
-
-        if self.people.t % self.resfreq == self.resfreq - 1:
-            # Update stock and flows
-            idx = int(self.people.t / self.resfreq)
-            self.results['hiv_infections'][idx] = self.people.scale_flows(hiv_inds)
-            self.results['hiv_infections_by_age'][:, idx] = np.histogram(self.people.age[hiv_inds], bins=self.people.age_bins, weights=self.people.scale[hiv_inds])[0]
-            self.results['n_hiv'][idx] = self.people.count('hiv')
-            hivinds = hpu.true(self.people['hiv'])
-            self.results['n_hiv_by_age'][:, idx] = np.histogram(self.people.age[hivinds], bins=self.people.age_bins, weights=self.people.scale[hivinds])[0]
 
         return hiv_inds
 
@@ -219,9 +336,54 @@ class HIVsim(hpb.ParsObj):
             gpars = self.people.pars['genotype_pars'][self.people.pars['genotype_map'][g]]
             hpv_inds = hpu.itruei((self.people.is_female & self.people.episomal[g, :]), hiv_inds)  # Women with HIV who have episomal HPV
             if len(hpv_inds):  # Reevaluate these women's severity markers and determine whether they will develop cellular changes
+<<<<<<< HEAD
                 self.people.set_severity(hpv_inds, g, gpars, dt)
+=======
+                self.people.set_severity_pars(hpv_inds, g, gpars)
+                self.people.set_severity(hpv_inds, g, gpars, dt, set_sev=False)
+>>>>>>> hiv-updates
         return
 
+    def update_hiv_results(self, hiv_inds):
+        if self.people.t % self.resfreq == self.resfreq - 1:
+            # Update stock and flows
+            idx = int(self.people.t / self.resfreq)
+            self.results['hiv_infections'][idx] = self.people.scale_flows(hiv_inds)
+            self.results['hiv_infections_by_age'][:, idx] = np.histogram(self.people.age[hiv_inds], bins=self.people.age_bins, weights=self.people.scale[hiv_inds])[0]
+            self.results['n_hiv'][idx] = self.people.count('hiv')
+            hivinds = hpu.true(self.people['hiv'])
+            self.results['n_hiv_by_age'][:, idx] = np.histogram(self.people.age[hivinds], bins=self.people.age_bins, weights=self.people.scale[hivinds])[0]
+
+            # Pull out those on ART:
+            self.results['n_art'][idx] = self.people.count('art')
+
+            # Pull out those with HPV and HIV+
+            hpvhivinds = hpu.true((self.people['hiv']) & self.people['infectious'])
+            self.results['n_hpv_by_age_with_hiv'][:, idx] = np.histogram(self.people.age[hpvhivinds], bins=self.people.age_bins, weights=self.people.scale[hpvhivinds])[0]
+
+            # Pull out those with HPV and HIV-
+            hpvnohivinds = hpu.true(~(self.people['hiv']) & self.people['infectious'])
+            self.results['n_hpv_by_age_no_hiv'][:, idx] = np.histogram(self.people.age[hpvnohivinds], bins=self.people.age_bins, weights=self.people.scale[hpvnohivinds])[0]
+
+            # Pull out those with cancer and HIV+
+            cancer_today_inds = hpu.true(self.people.date_cancerous == self.people.t)
+            if len(cancer_today_inds):
+                hiv_bools = self.people.hiv[cancer_today_inds]
+                cancer_today_hiv_pos_inds = cancer_today_inds[hiv_bools]
+                cancer_today_hiv_neg_inds = cancer_today_inds[~hiv_bools]
+                self.results['cancers_with_hiv'][idx] = self.people.scale_flows(cancer_today_hiv_pos_inds)
+                self.results['cancers_no_hiv'][idx] = self.people.scale_flows(cancer_today_hiv_neg_inds)
+                self.results['cancers_by_age_with_hiv'][:, idx] = np.histogram(self.people.age[cancer_today_hiv_pos_inds], bins=self.people.age_bins, weights=self.people.scale[cancer_today_hiv_pos_inds])[0]
+                self.results['cancers_by_age_no_hiv'][:, idx] = np.histogram(self.people.age[cancer_today_hiv_neg_inds], bins=self.people.age_bins, weights=self.people.scale[cancer_today_hiv_neg_inds])[0]
+
+            alive_female_hiv_inds = hpu.true(self.people.alive*self.people.is_female*self.people.hiv)
+            self.results['n_females_with_hiv_alive'][idx] = self.people.scale_flows(alive_female_hiv_inds)
+            self.results['n_females_with_hiv_alive_by_age'][:, idx] = np.histogram(self.people.age[alive_female_hiv_inds], bins=self.people.age_bins,
+                 weights=self.people.scale[alive_female_hiv_inds])[0]
+            alive_female_no_hiv_inds = hpu.true(self.people.alive * self.people.is_female * ~self.people.hiv)
+            self.results['n_females_no_hiv_alive'][idx] = self.people.scale_flows(alive_female_no_hiv_inds)
+            self.results['n_females_no_hiv_alive_by_age'][:, idx] = np.histogram(self.people.age[alive_female_no_hiv_inds], bins=self.people.age_bins,
+                 weights=self.people.scale[alive_female_no_hiv_inds])[0]
 
     def get_hiv_data(self, hiv_datafile=None, art_datafile=None):
         '''
@@ -291,6 +453,7 @@ class HIVsim(hpb.ParsObj):
         res = self.results
         simres = sim.results
 
+
         # Compute HIV incidence and prevalence
         def safedivide(num, denom):
             ''' Define a variation on sc.safedivide that respects shape of numerator '''
@@ -302,8 +465,20 @@ class HIVsim(hpb.ParsObj):
                 answer[:, fill_inds] = num[:, fill_inds] / denom[fill_inds]
             return answer
 
+        ng = sim.pars['n_genotypes']
+        no_hiv_by_age = simres['n_alive_by_age'][:] - res['n_hiv_by_age'][:]
         self.results['hiv_prevalence_by_age'][:] = safedivide(res['n_hiv_by_age'][:], simres['n_alive_by_age'][:])
         self.results['hiv_incidence'][:] = sc.safedivide(res['hiv_infections'][:], (simres['n_alive'][:] - res['n_hiv'][:]))
         self.results['hiv_incidence_by_age'][:] = sc.safedivide(res['hiv_infections_by_age'][:], (simres['n_alive_by_age'][:] - res['n_hiv_by_age'][:]))
-        self.results['hiv_prevalence'][:] = sc.safedivide(res['n_hiv'][:], simres['n_alive'][:])
+        self.results['hiv_prevalence'][:] = res['n_hiv'][:]/ simres['n_alive'][:]
+        self.results['hpv_prevalence_by_age_with_hiv'][:] = safedivide(res['n_hpv_by_age_with_hiv'][:], ng*res['n_hiv_by_age'][:])
+        self.results['hpv_prevalence_by_age_no_hiv'][:] = safedivide(res['n_hpv_by_age_no_hiv'][:], ng*no_hiv_by_age)
+        self.results['art_coverage'][:] = res['n_art'][:]/res['n_hiv'][:]
+
+        # Compute cancer incidence
+        scale_factor = 1e5  # Cancer incidence are displayed as rates per 100k women
+        self.results['cancer_incidence_with_hiv'][:] = res['cancers_with_hiv'][:]/ res['n_females_with_hiv_alive'][:]*scale_factor
+        self.results['cancer_incidence_no_hiv'][:] = res['cancers_no_hiv'][:]/ res['n_females_no_hiv_alive'][:]*scale_factor
+
+        sim.results = sc.mergedicts(simres, self.results)
         return