+++ hpvsim/sim.py
@@ -25,7 +25,7 @@ class Sim(hpb.BaseSim):
 
     def __init__(self, pars=None, datafile=None, label=None,
                  popfile=None, people=None, version=None, hiv_datafile=None, art_datafile=None,
-                 hiv_pars=None, **kwargs):
+                 **kwargs):
 
         # Set attributes
         self.label         = label    # The label/name of the simulation
@@ -33,7 +33,6 @@ class Sim(hpb.BaseSim):
         self.datafile      = datafile # The name of the data file
         self.art_datafile  = art_datafile # The name of the ART data file
         self.hiv_datafile  = hiv_datafile # The name of the HIV data file
-        self.hiv_pars      = hiv_pars
         self.popfile       = popfile  # The population file
         self.data          = None     # The data
         self.popdict       = people   # The population dictionary
@@ -495,8 +494,6 @@ class Sim(hpb.BaseSim):
         results['n_females_alive_by_age'] = init_res('Number females alive by age', n_rows=na)
         results['cdr'] = init_res('Crude death rate', scale=False)
         results['cbr'] = init_res('Crude birth rate', scale=False, color='#fcba03')
-        results['hiv_incidence'] = init_res('HIV incidence')
-        results['hiv_prevalence'] = init_res('HIV prevalence')
         results['hpv_prevalence'] = init_res('HPV prevalence', color=stock_colors[0])
         results['hpv_prevalence_by_genotype'] = init_res('HPV prevalence', n_rows=ng, color=stock_colors[0])
         results['hpv_prevalence_by_age'] = init_res('HPV prevalence by age', n_rows=na, color=stock_colors[0])
@@ -571,7 +568,7 @@ class Sim(hpb.BaseSim):
         
         # Finish initialization
         self.hivsim = hphiv.HIVsim(self, hiv_datafile=self.hiv_datafile, art_datafile=self.art_datafile,
-                                   hiv_pars=self.hiv_pars)
+                                   hiv_pars=self['hiv_pars'])
         self.people.initialize(sim_pars=self.pars, hivsim=self.hivsim) # Fully initialize the people
         self.reset_layer_pars(force=False) # Ensure that layer keys match the loaded population
         if init_states:
@@ -851,9 +848,6 @@ class Sim(hpb.BaseSim):
             for key in [state.name for state in hpd.PeopleMeta.intv_states]:
                 self.results[f'n_{key}'][idx] = people.count(key)
 
-            # Count total hiv infections
-            self.results['n_hiv'][idx] = people.count('hiv')
-
             # Update cancers and cancers by age
             cases_by_age = self.results['cancers_by_age'][:, idx]
             inds = people.alive * (self.people.sex==0) * ~people.cancerous.any(axis=0)
@@ -1028,8 +1022,6 @@ class Sim(hpb.BaseSim):
         self.results['hpv_prevalence'][:]               = sc.safedivide(res['n_infectious'][:], ng*res['n_alive'][:])
         self.results['hpv_prevalence_by_genotype'][:]   = safedivide(res['n_infectious_by_genotype'][:], res['n_alive'][:])
         self.results['hpv_prevalence_by_age'][:]        = safedivide(res['n_infectious_by_age'][:], ng*res['n_alive_by_age'][:])
-        self.results['hiv_incidence'][:]                = sc.safedivide(res['hiv_infections'][:], (res['n_alive'][:]-res['n_hiv'][:]))
-        self.results['hiv_prevalence'][:]               = sc.safedivide(res['n_hiv'][:], res['n_alive'][:])
 
         alive_females = res['n_alive_by_sex'][0,:]
 
@@ -1049,13 +1041,13 @@ class Sim(hpb.BaseSim):
         self.results['cin3_prevalence_by_genotype'][:] = safedivide(res['n_cin3_by_genotype'][:], alive_females)
         self.results['cin3_prevalence_by_age'][:] = safedivide(res['n_cin3_by_age'][:],
                                                                ng*res['n_females_alive_by_age'][:])
-        # Compute CIN and cancer incidence.
+        # Compute cancer incidence.
         at_risk_females = alive_females - res['n_cancerous'][:]
-        scale_factor = 1e5  # Cancer and CIN incidence are displayed as rates per 100k women
+        scale_factor = 1e5  # Cancer incidence are displayed as rates per 100k women
         demoninator = at_risk_females / scale_factor
         self.results['cancer_incidence'][:]             = res['cancers'][:] / demoninator
         self.results['cancer_incidence_by_genotype'][:] = res['cancers_by_genotype'][:] / demoninator
-        self.results['cancer_incidence_by_age'][:]      = sc.safedivide(res['cancers'][:], res['n_females_alive_by_age'][:]/scale_factor)
+        self.results['cancer_incidence_by_age'][:]      = sc.safedivide(res['cancers_by_age'][:], res['n_females_alive_by_age'][:]/scale_factor)
 
         # Compute cancer mortality. Denominator is all women alive
         denominator = alive_females/scale_factor