+++ tests/devtests/test_hiv.py
@@ -5,31 +5,97 @@ Tests for single simulations
 #%% Imports and settings
 import sciris as sc
 import hpvsim as hpv
+import numpy as np
 
 do_plot = 0
 do_save = 0
-debug = 1
+debug = 0
 
 n_agents = [50e3,500][debug] # Swap between sizes
+start = [1950,1990][debug]
+ms_agent_ratio = [100,10][debug]
 
 
 #%% Define the tests
 
-def test_hiv(model_hiv=True):
+def test_calibration_hiv():
+    sc.heading('Testing calibration with hiv pars')
+    pars = {
+        'n_agents': n_agents,
+        'location': 'south africa',
+        'model_hiv': True,
+        'start': start,
+        'end': 2020,
+    }
+
+    hiv_datafile='hiv_incidence_south_africa.csv'
+    art_datafile='art_coverage_south_africa.csv'
+
+    sim = hpv.Sim(
+        pars=pars,
+        hiv_datafile=hiv_datafile,
+        art_datafile=art_datafile
+    )
+
+    calib_pars = dict(
+        beta=[0.05, 0.010, 0.20],
+        dur_transformed=dict(par1=[5, 3, 10]),
+    )
+    genotype_pars = dict(
+        hpv16=dict(
+            sev_rate=[0.5, 0.2, 1.0],
+        ),
+        hpv18=dict(
+            sev_rate=[0.5, 0.2, 1.0],
+        )
+    )
+
+    hiv_pars = dict(
+        rel_sus= dict(
+            cat1=dict(value=[3, 2,4])
+        )
+    )
+
+
+    calib = hpv.Calibration(sim, calib_pars=calib_pars, genotype_pars=genotype_pars, hiv_pars=hiv_pars,
+                            datafiles=[
+                                'south_africa_hpv_data.csv',
+                                'south_africa_cancer_data_2020.csv',
+                                'south_africa_cancer_data_hiv_2020.csv',
+                            ],
+                            total_trials=3, n_workers=1)
+    calib.calibrate(die=True)
+    calib.plot(res_to_plot=4)
+    return sim, calib
+
+
+def test_hiv():
     sc.heading('Testing hiv')
+
+    partners = dict(m=dict(dist='poisson', par1=0.1),
+                    c=dict(dist='poisson', par1=0.5),
+                    o=dict(dist='poisson', par1=0.0),
+                    )
 
     pars = {
         'n_agents': n_agents,
         'location': 'south africa',
-        'model_hiv': model_hiv
+        'model_hiv': True,
+        'start': start,
+        'end': 2020,
+        'ms_agent_ratio': ms_agent_ratio,
+        'partners': partners,
+        'cross_layer': 0.1  # Proportion of females who have crosslayer relationships
+        # 'hiv_pars': {
+        # 'rel_sus': dict(
+        #     cat1=dict(value=3)
+        # )
+        # }
     }
 
-    if model_hiv:
-        hiv_datafile='./test_data/hiv_incidence_south_africa.csv'
-        art_datafile = './test_data/art_coverage_south_africa.csv'
-    else:
-        hiv_datafile=None
-        art_datafile=None
+
+    hiv_datafile = '../test_data/hiv_incidence_south_africa.csv'
+    art_datafile = '../test_data/art_coverage_south_africa.csv'
 
     sim = hpv.Sim(
         pars=pars,
@@ -37,7 +103,26 @@ def test_hiv(model_hiv=True):
         art_datafile=art_datafile
     )
     sim.run()
-    sim.plot(to_plot=['hiv_prevalence'])
+    to_plot = {
+        'ART Coverage': [
+            'art_coverage',
+        ],
+        'HPV prevalence by HIV status': [
+            'hpv_prevalence_by_age_with_hiv',
+            'hpv_prevalence_by_age_no_hiv'
+        ],
+        'Age standardized cancer incidence (per 100,000 women)': [
+            'asr_cancer_incidence',
+            'cancer_incidence_with_hiv',
+            'cancer_incidence_no_hiv',
+        ],
+        'Cancers by age and HIV status': [
+            'cancers_by_age_with_hiv',
+            'cancers_by_age_no_hiv'
+        ]
+    }
+    sim.plot()
+    sim.plot(to_plot=to_plot)
     return sim
 
 
@@ -47,6 +132,8 @@ def test_impact_on_cancer():
     pars = {
         'n_agents': n_agents,
         'location': 'south africa',
+        'start': start,
+        'end': 2030
     }
 
     base_sim = hpv.Sim(
@@ -63,19 +150,9 @@ def test_impact_on_cancer():
             }
         },
         'hiv_baseline': {
-            'name': 'HIV, baseline',
+            'name': 'HIV',
             'pars': {
                 'model_hiv': True
-            }
-        },
-        'hiv_elevated_risk': {
-            'name': 'HIV, elevated risk',
-            'pars': {
-                'model_hiv': True,
-            },
-            'hiv_pars': {
-                'rel_sus': 3,
-                'rel_hiv_sev_infl': {'cd4_200': 0.36, 'cd4_200_500': 0.76},
             }
         }
     }
@@ -84,15 +161,13 @@ def test_impact_on_cancer():
     scens = hpv.Scenarios(sim=base_sim, metapars=metapars, scenarios=scenarios)
     scens.run(debug=debug)
     to_plot = {
-        'HIV prevalence': [
-            'hiv_prevalence',
-        ],
         'HPV prevalence': [
             'hpv_prevalence',
         ],
         'Age standardized cancer incidence (per 100,000 women)': [
             'asr_cancer_incidence',
-        ]
+        ],
+
     }
     scens.plot(to_plot=to_plot)
     return scens
@@ -102,7 +177,8 @@ if __name__ == '__main__':
 
     # Start timing and optionally enable interactive plotting
     T = sc.tic()
-    sim0 = test_hiv(model_hiv=True)
+    sim0 = test_hiv()
     # sim1 = test_impact_on_cancer()
+    # sim, calib = test_calibration_hiv()
     sc.toc(T)
     print('Done.')